requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
boto==2.49.0
boto3==1.9.253
botocore==1.12.253
certifi==2020.4.5.1
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
decorator==4.4.0
docutils==0.15.2
dogpile.cache==0.7.1
idna==2.8
iso8601==0.1.12
jmespath==0.9.4
jsonpatch==1.23
jsonpointer==2.0
kafka-python==2.0.1
keystoneauth1==3.13.1
munch==2.3.2
netifaces==0.10.9
openstacksdk==0.26.0
os-service-types==1.6.0
pbr==5.1.3
pycparser==2.19
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
requestsexceptions==1.4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
six==1.12.0
stevedore==1.30.1
urllib3==1.24.1
